refactor(components): make showcase list components interactive
