+++ src/containers/menu/components/showcase/button.component.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import {
+  Button,
+  ButtonProps,
+} from '@kitten/ui';
+
+export const ButtonShowcase = (props?: ButtonProps): React.ReactElement<ButtonProps> => {
+  return (
+    <Button
+      {...props}>
+      BUTTON
+    </Button>
+  );
+};
