+++ src/containers/menu/components/showcase/buttonGroup.component.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import {
+  Button,
+  ButtonGroup,
+  ButtonGroupProps,
+} from '@kitten/ui';
+import { StarIcon } from '@src/assets/icons';
+
+export const ButtonGroupShowcase = (props?: ButtonGroupProps): React.ReactElement<ButtonGroupProps> => {
+  return (
+    <ButtonGroup {...props}>
+      <Button icon={StarIcon}/>
+      <Button icon={StarIcon}/>
+    </ButtonGroup>
+  );
+};
