+++ src/containers/menu/components/showcase/complex.component.tsx
@@ -1,0 +1,16 @@
+import {
+  ImageStyle,
+  StyleProp,
+} from 'react-native';
+import { theme } from 'eva/packages/theme/eva';
+import { StarIcon } from '@src/assets/icons';
+
+export const ComplexComponentShowcase = () => {
+  const style: StyleProp<ImageStyle> = {
+    width: 32,
+    height: 32,
+    tintColor: theme['color-primary-500'],
+  };
+
+  return StarIcon(style);
+};
