+++ src/containers/menu/components/showcase/input.component.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import {
+  Input,
+  InputProps,
+} from '@kitten/ui';
+
+export const InputShowcase = (props?: InputProps): React.ReactElement<InputProps> => {
+  return (
+    <Input
+      {...props}
+      placeholder='Text'
+    />
+  );
+};
